+++ spec/unit/chef/knife/container_docker_init_spec.rb
@@ -17,6 +17,7 @@
 
 require 'spec_helper'
 require 'chef/knife/container_docker_init'
+Chef::Knife::ContainerDockerInit.load_deps
 
 describe Chef::Knife::ContainerDockerInit do
 
@@ -27,6 +28,10 @@ describe Chef::Knife::ContainerDockerInit do
     stdout_io.string
   end
 
+  def generator_context
+    KnifeContainer::Generator.context
+  end
+
   let(:default_cookbook_path) do
     File.expand_path('cookbooks', fixtures_path)
   end
@@ -34,8 +39,6 @@ describe Chef::Knife::ContainerDockerInit do
   let(:reset_chef_config) do
     Chef::Config.reset
     Chef::Config[:chef_repo_path] = tempdir
-    Chef::Config[:knife][:dockerfiles_path] = File.join(tempdir, 'dockerfiles')
-    Chef::Config[:knife][:docker_image] = nil
     Chef::Config[:cookbook_path] = File.join(fixtures_path, 'cookbooks')
     Chef::Config[:chef_server_url] = "http://localhost:4000"
     Chef::Config[:validation_key] = File.join(fixtures_path, '.chef','validator.pem')
@@ -44,88 +47,75 @@ describe Chef::Knife::ContainerDockerInit do
     Chef::Config[:encrypted_data_bag_secret] = File.join(fixtures_path, '.chef', 'encrypted_data_bag_secret')
   end
 
-  def generator_context
-    KnifeContainer::Generator.context
-  end
-
   subject(:knife) do
     Chef::Knife::ContainerDockerInit.new(argv).tap do |c|
       allow(c).to receive(:output).and_return(true)
       allow(c.ui).to receive(:stdout).and_return(stdout_io)
       allow(c.chef_runner).to receive(:stdout).and_return(stdout_io)
-      c.parse_options(argv)
-      c.merge_configs
-      c.set_config_defaults
     end
   end
 
   let(:argv) { %w[ docker/demo ] }
-  let(:dockercontext_path) { "#{Chef::Config[:dockerfiles_path]}/docker/demo" }
+  let(:dockercontext_path) { "#{Chef::Config[:knife][:dockerfiles_path]}/docker/demo" }
 
-  before(:each) { reset_chef_config }
+  it 'initializes a new docker context' do
+    allow(knife).to receive(:docker_context_path).and_return(dockercontext_path)
+    expect(knife).to receive(:validate!)
+    expect(knife).to receive(:setup_context)
+    expect(knife.chef_runner).to receive(:converge)
+    expect(knife).to receive(:download_and_tag_base_image)
+    expect(knife.ui).to receive(:info)
+    knife.run
+  end
 
-  describe '#run' do
-    before do
-      allow(knife).to receive(:docker_context_path).and_return()
-    end
+  describe 'when argv is empty' do
+    let(:argv) { %W[] }
 
-    it 'initializes a new docker context' do
-      expect(knife).to receive(:set_config_defaults)
-      expect(knife).to receive(:validate)
-      expect(knife).to receive(:setup_context)
-      expect(knife.chef_runner).to receive(:converge)
-      expect(knife).to receive(:download_and_tag_base_image)
-      knife.run
-    end
-
-    context 'when argv is empty' do
-      let(:argv) { %W[] }
-
-      it 'throws an error and prints a message' do
-        lambda do
-          expect(knife).to receive(:show_usage)
-          expect(knife.ui).to receive(:fatal)
-          knife.run
-          expect(knife).to have_received(:exit).with(false)
-        end
+    it 'throws an error and prints a message' do
+      lambda do
+        expect(knife).to receive(:error_out)
+        knife.run
       end
     end
   end
 
-  describe 'when no cli overrides have been specified' do
-    it 'sets values to Chef::Config default values' do
-      expect(knife.config[:chef_server_url]).to eq('http://localhost:4000')
-      expect(knife.config[:cookbook_path]).to eq(File.join(fixtures_path, 'cookbooks'))
-      expect(knife.config[:node_path]).to eq(File.join(tempdir, 'nodes'))
-      expect(knife.config[:role_path]).to eq(File.join(tempdir, 'roles'))
-      expect(knife.config[:environment_path]).to eq(File.join(tempdir, 'environments'))
-      expect(knife.config[:data_bag_path]).to eq(File.join(tempdir, 'data_bags'))
-      expect(knife.config[:dockerfiles_path]).to eq(File.join(tempdir, 'dockerfiles'))
-      expect(knife.config[:run_list]).to eq([])
+  describe 'when establishing configuration' do
+    before(:each) do
+      reset_chef_config
+      allow(knife).to receive(:validate!) # skip validation
+      allow(knife.chef_runner).to receive(:converge) # skip converge
+      allow(knife).to receive(:download_and_tag_base_image) # skip image DL
+      allow(knife.ui).to receive(:info) # skip printing the output
     end
 
-    context 'and cookbook_path is an array' do
-      before do
-        Chef::Config[:cookbook_path] = [
-          '/path/to/cookbooks',
-          '/path/to/site-cookbooks'
-        ]
-      end
+    subject(:config) { knife.config }
 
-      it 'honors the array' do
-        expect(knife.config[:cookbook_path]).to eq([
-          '/path/to/cookbooks',
-          '/path/to/site-cookbooks'
-        ])
+    describe 'when no cli overrides have been specified' do
+      it 'sets values to Chef::Config default values' do
+        knife.run
+        expect(config[:base_image]).to eql('chef/ubuntu-12.04:latest')
+        expect(config[:dockerfiles_path]).to eql(File.join(tempdir, 'dockerfiles'))
+        expect(config[:berksfile_source]).to eql('https://supermarket.getchef.com')
       end
     end
 
-    context 'when base image is specified' do
+    describe 'when overrides are specified' do
+      context 'using Chef::Config' do
+        it 'respects that value' do
+          Chef::Config[:knife][:base_docker_image] = 'docker/demo:11.12.8'
+          Chef::Config[:knife][:berksfile_source] = 'https://mychefserver.com'
+          knife.run
+          expect(config[:base_image]).to eql('docker/demo:11.12.8')
+          expect(config[:berksfile_source]).to eql('https://mychefserver.com')
+        end
+      end
+
       context 'with a tag' do
         let(:argv) { %w[ docker/demo -f docker/demo:11.12.8 ] }
 
         it 'respects that tag' do
-          expect(knife.config[:base_image]).to eql('docker/demo:11.12.8')
+          knife.run
+          expect(config[:base_image]).to eql('docker/demo:11.12.8')
         end
       end
 
@@ -133,28 +123,27 @@ describe Chef::Knife::ContainerDockerInit do
         let(:argv) { %w[ docker/demo -f docker/demo ] }
 
         it 'should append the \'latest\' tag on the name' do
-          expect(knife.config[:base_image]).to eql("docker/demo:latest")
+          knife.run
+          expect(config[:base_image]).to eql("docker/demo:latest")
         end
       end
     end
   end
 
   #
-  # The chef runner converge
+  # Building out the context
   #
   describe 'when creating the context' do
-
     before do
-      allow(knife).to receive(:download_and_tag_base_image) # dont download
-      allow(knife).to receive(:setup_and_verify_docker) # dont validate docker
-      allow(knife).to receive(:verify_docker_context) # dont validate docker
-      allow(knife).to receive(:docker_context_name).and_return('docker/demo')
-      allow(knife).to receive(:docker_context_path).and_return("#{tempdir}/dockerfiles/docker/demo")
+      reset_chef_config
+      allow(knife).to receive(:validate!) # skip validation
+      allow(knife).to receive(:download_and_tag_base_image) # skip image DL
+      allow(knife.ui).to receive(:info) # skip printing the output
     end
 
     context 'when -b is passed' do
-      before(:each) { reset_chef_config }
-      let(:argv) { %W[
+      before { reset_chef_config }
+      let(:argv) { %w[
         docker/demo
         -r role[demo],recipe[demo::recipe],recipe[nginx]
         -b
@@ -172,7 +161,6 @@ describe Chef::Knife::ContainerDockerInit do
     end
 
     context 'when -z is passed' do
-      before { reset_chef_config }
       let(:argv) { %w[ docker/demo -z ] }
 
       let(:expected_container_file_relpaths) do
@@ -211,9 +199,8 @@ describe Chef::Knife::ContainerDockerInit do
     end
 
     describe 'when copying cookbooks to temporary chef-repo' do
-      let(:argv) { %W[docker/demo -r recipe[nginx],recipe[apt] ]}
+      let(:argv) { %W[docker/demo -r recipe[nginx],recipe[apt] -z ]}
       before do
-        reset_chef_config
         Chef::Config[:cookbook_path] = [
           "#{fixtures_path}/cookbooks",
           "#{fixtures_path}/site-cookbooks"
@@ -232,21 +219,16 @@ describe Chef::Knife::ContainerDockerInit do
     end
 
     describe 'when invalid cookbook path is specified' do
-      let(:argv) { %W[ docker/demo -r recipe[nginx] ]}
-
-      before(:each) do
-        reset_chef_config
-        Chef::Config[:cookbook_path] = '/tmp/nil/cookbooks'
-      end
+      let(:argv) { %W[ docker/demo -r recipe[nginx] -z]}
 
       it 'logs an error and does not copy cookbooks' do
+        Chef::Config[:cookbook_path] = '/tmp/nil/cookbooks'
         knife.run
         expect(stdout).to include('log[Could not find a \'/tmp/nil/cookbooks\' directory in your chef-repo.] action write')
       end
     end
 
     describe 'when creating chef/client.rb' do
-      before { reset_chef_config }
       subject(:config_file) do
         File.read("#{tempdir}/dockerfiles/docker/demo/chef/client.rb")
       end
@@ -265,12 +247,8 @@ describe Chef::Knife::ContainerDockerInit do
       end
 
       context 'and the encrypted_data_bag_secret is not specified' do
-        before do
-          reset_chef_config
-          Chef::Config[:encrypted_data_bag_secret] = nil
-        end
-
         it 'does not add encrypted_data_bag_secret value to config' do
+          Chef::Config[:encrypted_data_bag_secret] = nil
           knife.run
           expect(config_file).to_not include 'encrypted_data_bag_secret'
         end
@@ -283,10 +261,7 @@ describe Chef::Knife::ContainerDockerInit do
         File.read("#{Chef::Config[:chef_repo_path]}/dockerfiles/docker/demo/Dockerfile")
       end
 
-      before do
-        reset_chef_config
-        knife.run
-      end
+      before { knife.run }
 
       it 'sets the base_image name in a comment in the Dockerfile' do
         expect(dockerfile).to include '# BASE chef/ubuntu-12.04:latest'